# src/app/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  controller.CatItem:
    properties:
      adoption_price:
        type: integer
      cat_thumbnail_path:
        type: string
      contract_benefit:
        type: integer
      contract_days:
        type: integer
      deposit:
        type: integer
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      pet_coin:
        type: integer
      price:
        type: integer
      reservation_price:
        type: integer
      status:
        type: integer
    type: object
...
  controller.GetTimeScheduleCatResponse:
    properties:
      cats:
        items:
          $ref: '#/definitions/controller.CatItem'
        type: array
      lower:
        type: integer
      total:
        type: integer
      upper:
        type: integer
    type: object
  controller.GetUserInfoResponse:
    properties:
      identified_code:
        type: string
      phone:
        type: string
      user_name:
        type: string
    type: object
  controller.GetUserListResponse:
    properties:
      lower:
        type: integer
      total:
        type: integer
      upper:
        type: integer
      users:
        items:
          $ref: '#/definitions/controller.UserItem'
        type: array
    type: object
  controller.Message:
    properties:
      message:
        example: message
        type: string
    type: object
...
  controller.UserItem:
    properties:
      id:
        type: integer
      identified_code:
        type: string
      phone:
        type: string
      user_name:
        type: string
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
...
  /api/v1/cats/time/schedules/{scheduleId}:
    get:
      consumes:
      - application/json
      description: get cat list with timeSchedule from database
      parameters:
      - description: 時段的ID
        in: path
        name: scheduleId
        required: true
        type: integer
      - description: 貓列表的lower
        in: query
        name: lower
        required: true
        type: integer
      - description: 貓列表的upper
        in: query
        name: upper
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetTimeScheduleCatResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
  /api/v1/time/schedules/{scheduleId}/cat/{catId}:
    post:
      consumes:
      - application/json
      description: Add a new timePeriodCatPivot
      parameters:
      - description: 貓的ID
        in: path
        name: catId
        required: true
        type: integer
      - description: 時段的ID
        in: path
        name: scheduleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
  /api/v1/user/info:
    get:
      consumes:
      - application/json
      description: get user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetUserInfoResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/user/login:
    post:
      consumes:
      - multipart/form-data
      description: user account login
      parameters:
      - description: 用戶手機號碼
        in: formData
        name: phone
        required: true
        type: string
      - description: 用戶密碼
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/user/logout:
    post:
      consumes:
      - multipart/form-data
      description: user account logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/user/register:
    post:
      consumes:
      - multipart/form-data
      description: user account register
      parameters:
      - description: 用戶手機號碼
        in: formData
        name: phone
        required: true
        type: string
      - description: 用戶手機認證碼
        in: formData
        name: valid_value
        required: true
        type: string
      - description: 用戶名稱
        in: formData
        name: user_name
        required: true
        type: string
      - description: 用戶密碼
        in: formData
        name: password
        required: true
        type: string
      - description: 用戶二級密碼
        in: formData
        name: security_password
        required: true
        type: string
      - description: 推薦用戶識別碼
        in: formData
        name: identified_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: get user list
      parameters:
      - description: 用戶列表的lower
        in: query
        name: lower
        required: true
        type: integer
      - description: 用戶列表的upper
        in: query
        name: upper
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetUserListResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
swagger: "2.0"